@@ -1,7 +1,19 @@
-import { TextField, Button, Box } from "@material-ui/core";
-import { FC } from "react";
+import {
+  Box,
+  TextField,
+  Avatar,
+  Typography,
+  List,
+  ListItem,
+  ListItemAvatar,
+  ListItemText,
+  ListItemButton,
+} from "@mui/material";
+import dayjs from "dayjs";
+import React, { FC } from "react";
 import { Controller } from "react-hook-form";
 import { FormViewProps, IssueFormFields } from "./IssueForm.types";
+import CalculateIcon from "@mui/icons-material/Calculate";
 
 export const FormView: FC<FormViewProps> = ({
   onSubmit,
@@ -13,73 +25,78 @@ export const FormView: FC<FormViewProps> = ({
   <form onSubmit={handleSubmit(onSubmit)}>
     <Box>
       <Controller
-        name={IssueFormFields.Owner}
-        control={control}
-        render={({ field }) => (
-          <TextField {...field} variant="filled" fullWidth label={"Owner"} />
-        )}
-      />
-    </Box>
-    <Box>
-      <Controller
-        name={IssueFormFields.Repo}
-        control={control}
-        render={({ field }) => (
-          <TextField
-            {...field}
-            variant="filled"
-            fullWidth
-            label={"Repository"}
-          />
-        )}
-      />
-    </Box>
-    <Box>
-      <Controller
         name={IssueFormFields.Title}
         control={control}
         render={({ field }) => (
-          <TextField {...field} variant="filled" fullWidth label={"Title"} />
+          <TextField {...field} variant="outlined" fullWidth label={"Title"} />
         )}
       />
     </Box>
-    <Box>
+    <Box mt={2}>
       <Controller
         name={IssueFormFields.Body}
         control={control}
         render={({ field }) => (
           <TextField
             {...field}
-            variant="filled"
+            variant="outlined"
             fullWidth
             label={"Body"}
             multiline
+            rows={10}
           />
         )}
       />
     </Box>
-    <Box>
-      <Button type={"submit"} fullWidth>
-        submit
-      </Button>
-    </Box>
-    {isLoading && <Box>Loading...</Box>}
 
     {data && (
-      <Box>
-        labels:
-        <ul>
-          {data.labels.map(({ id, name }) => (
-            <li key={id}>{name}</li>
-          ))}
-        </ul>
-        participants:
-        <ul>
-          {data.participants.map((name) => (
-            <li key={name}>{name}</li>
-          ))}
-        </ul>
-        {data.closedAt && <Box>closed at: {data.closedAt}</Box>}
+      <Box mt={10}>
+        {isLoading && <Box>Loading...</Box>}
+
+        <Box>
+          {data.labels?.length > 0 && (
+            <Box>
+              <Typography>{`Estimated labels:`}</Typography>
+              <ul>
+                {data.labels.map(({ id, name }) => (
+                  <li key={id}>{name}</li>
+                ))}
+              </ul>
+            </Box>
+          )}
+
+          {data.participants?.length > 0 && (
+            <Box>
+              <Typography variant="h5">{`Users can help you`}</Typography>
+              <List>
+                {data.participants.map(({ login, name, avatarUrl }) => (
+                  <ListItem key={login}>
+                    <ListItemButton
+                      component="a"
+                      href={`https://github.com/${login}`}
+                      target="_blank"
+                    >
+                      <ListItemAvatar>
+                        <Avatar src={avatarUrl} alt={login} />
+                      </ListItemAvatar>
+                      <ListItemText primary={name} />
+                    </ListItemButton>
+                  </ListItem>
+                ))}
+              </List>
+            </Box>
+          )}
+
+          {data.closedAt && (
+            <Box>
+              <Typography>
+                {`Estimated date solving the issue: ${dayjs(
+                  data.closedAt
+                ).format(`DD. MM. YYYY`)}`}
+              </Typography>
+            </Box>
+          )}
+        </Box>
       </Box>
     )}
   </form>
